ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.0.3
openai==1.3.0
transformers==4.35.0
torch==2.1.0



